update: add erc20_canister
