+++ tutorials/rust/http_outcall_icweb3/.env
@@ -2,7 +2,7 @@
 # DFX CANISTER ENVIRONMENT VARIABLES
 DFX_VERSION='0.13.1'
 DFX_NETWORK='local'
-CANISTER_ID_http_outcall_icweb3_backend='rrkah-fqaaa-aaaaa-aaaaq-cai'
+CANISTER_ID_erc20_canister='rrkah-fqaaa-aaaaa-aaaaq-cai'
 CANISTER_ID='rrkah-fqaaa-aaaaa-aaaaq-cai'
-CANISTER_CANDID_PATH='/Users/linnefromice/repository/github.com/_linnefromice/linnefromice/practice-icp/tutorials/rust/http_outcall_icweb3/src/http_outcall_icweb3_backend/http_outcall_icweb3_backend.did'
+CANISTER_CANDID_PATH='/Users/linnefromice/repository/github.com/_linnefromice/linnefromice/practice-icp/tutorials/rust/http_outcall_icweb3/src/erc20_canister/erc20_canister.did'
 # END DFX CANISTER ENVIRONMENT VARIABLES
