+++ tutorials/rust/http_outcall_icweb3/Cargo.toml
@@ -1,4 +1,4 @@
 [workspace]
 members = [
-    "src/http_outcall_icweb3_backend",
+    "src/erc20_canister",
 ]
