+++ tutorials/rust/http_outcall_icweb3/src/erc20_canister/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "http_outcall_icweb3_backend"
+name = "erc20_canister"
 version = "0.1.0"
 edition = "2021"
