+++ tutorials/rust/http_outcall_icweb3/src/erc20_canister/erc20_canister.did
@@ -1,0 +1,9 @@
+service : {
+    "get_settings": () -> (text) query;
+    "get_latest_block": () -> (text);
+    "call_name": (text) -> (text);
+    "call_symbol": (text) -> (text);
+    "call_decimals": (text) -> (nat8);
+    "call_total_supply": (text) -> (nat);
+    "call_balance_of": (text, text) -> (nat);
+}
