--- tutorials/rust/http_outcall_icweb3/src/http_outcall_icweb3_backend/http_outcall_icweb3_backend.did
@@ -1,5 +1,0 @@
-service : {
-    "greet": (text) -> (text) query;
-    "get_settings": () -> (text) query;
-    "get_latest_block": () -> (text);
-}
